streamlit_csv_editor/requirements.txt
streamlit
pandas
openpyxl
streamlit-aggrid
xlsxwriter
mysql-connector-python